refactor: 레포지토리 활용 로직을 응용 서비스로 이동
