+++ src/main/java/com/gdschongik/gdsc/domain/recruitment/application/AdminRecruitmentService.java
@@ -1,7 +1,6 @@
 package com.gdschongik.gdsc.domain.recruitment.application;
 
 import static com.gdschongik.gdsc.domain.common.model.SemesterType.*;
-import static com.gdschongik.gdsc.global.common.constant.TemporalConstant.*;
 import static com.gdschongik.gdsc.global.exception.ErrorCode.*;
 
 import com.gdschongik.gdsc.domain.common.model.SemesterType;
@@ -10,7 +9,6 @@ import com.gdschongik.gdsc.domain.recruitment.dao.RecruitmentRepository;
 import com.gdschongik.gdsc.domain.recruitment.dao.RecruitmentRoundRepository;
 import com.gdschongik.gdsc.domain.recruitment.domain.Recruitment;
 import com.gdschongik.gdsc.domain.recruitment.domain.RecruitmentRound;
-import com.gdschongik.gdsc.domain.recruitment.domain.RecruitmentValidator;
 import com.gdschongik.gdsc.domain.recruitment.domain.vo.Period;
 import com.gdschongik.gdsc.domain.recruitment.dto.request.RecruitmentCreateRequest;
 import com.gdschongik.gdsc.domain.recruitment.dto.request.RecruitmentRoundUpdateRequest;
@@ -31,11 +29,10 @@ public class AdminRecruitmentService {
 
     private final RecruitmentRepository recruitmentRepository;
     private final RecruitmentRoundRepository recruitmentRoundRepository;
-    private final RecruitmentValidator recruitmentValidator;
 
     @Transactional
     public void createRecruitment(RecruitmentCreateRequest request) {
-        recruitmentValidator.validateRecruitmentCreate(request.academicYear(), request.semesterType());
+        validateRecruitmentOverlap(request.academicYear(), request.semesterType());
 
         Recruitment recruitment = Recruitment.createRecruitment(
                 request.academicYear(),
@@ -70,6 +67,12 @@ public class AdminRecruitmentService {
         }
 
         recruitmentRounds.forEach(RecruitmentRound::validatePeriodNotStarted);
+    }
+
+    private void validateRecruitmentOverlap(Integer academicYear, SemesterType semesterType) {
+        if (recruitmentRepository.existsByAcademicYearAndSemesterType(academicYear, semesterType)) {
+            throw new CustomException(RECRUITMENT_OVERLAP);
+        }
     }
 
     private LocalDateTime getSemesterStartDate(Integer academicYear, SemesterType semesterType) {
